requirements.txt
analysis==0.1.7
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
bs4==0.0.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cssselect2==0.8.0
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.58.2
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfkit==1.0.0
pillow==11.2.1
preshed==3.0.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
textstat==0.7.3
thinc==8.3.6
tinycss2==1.4.0
tldextract==5.3.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
weasyprint==62.2
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xlsxwriter==3.2.5
zopfli==0.2.3.post1
psycopg2-binary